+++ examples/sumsquare_color.py
@@ -15,7 +15,7 @@ def sum_squares(imagePath, magnification=None, **kwargs):
     :param imagePath: path of the file to analyze.
     :param magnification: optional magnification to use for the analysis.
     """
-    source = large_image.getTileSource(imagePath)
+    source = large_image.open(imagePath)
 
     tileSumSquares = []
     # iterate through the tiles at a particular magnification: