+++ large_image/tilesource/base.py
@@ -1194,13 +1194,15 @@ class TileSource:
                 self.logger.exception('Failed to execute style function %s' % function['name'])
             return image
 
-    def getICCProfiles(self, idx=None):
+    def getICCProfiles(self, idx=None, onlyInfo=False):
         """
         Get a list of all ICC profiles that are available for the source, or
         get a specific profile.
 
         :param idx: a 0-based index into the profiles to get one profile, or
             None to get a list of all profiles.
+        :param onlyInfo: if idx is None and this is true, just return the
+            profile information.
         :returns: either one or a list of PIL.ImageCms.CmsProfile objects, or
             None if no profiles are available.  If a list, entries in the list
             may be None.
@@ -1218,6 +1220,10 @@ class TileSource:
             if idx == pidx:
                 return prof
             results.append(prof)
+        if onlyInfo:
+            results = [
+                PIL.ImageCms.getProfileInfo(prof).strip() or 'present'
+                if prof else None for prof in results]
         return results
 
     def _applyICCProfile(self, sc, frame):
@@ -1275,15 +1281,17 @@ class TileSource:
         sc = types.SimpleNamespace(
             image=image, originalStyle=style, x=x, y=y, z=z, frame=frame,
             mainImage=image, mainFrame=frame, dtype=None, axis=None)
-        if style is None:
-            sc.style = {'bands': []}
+        if style is None or ('icc' in style and len(style) == 1):
+            sc.style = {'icc': (style or {}).get(
+                'icc', config.getConfig('icc_correction', True)), 'bands': []}
         else:
             sc.style = style if 'bands' in style else {'bands': [style]}
             sc.dtype = style.get('dtype')
             sc.axis = style.get('axis')
-        if hasattr(self, '_iccprofiles') and sc.style.get('icc', True):
+        if hasattr(self, '_iccprofiles') and sc.style.get(
+                'icc', config.getConfig('icc_correction', True)):
             image = self._applyICCProfile(sc, frame)
-        if style is None or (len(style) == 1 and 'icc' in style):
+        if style is None or ('icc' in style and len(style) == 1):
             sc.output = image
         else:
             sc.output = numpy.zeros((image.shape[0], image.shape[1], 4), float)
@@ -1458,15 +1466,16 @@ class TileSource:
             maxX = (x + 1) * self.tileWidth
             maxY = (y + 1) * self.tileHeight
             isEdge = maxX > sizeX or maxY > sizeY
+        hasStyle = (
+            len(set(getattr(self, 'style', {})) - {'icc'}) or
+            getattr(self, 'style', {}).get('icc', config.getConfig('icc_correction', True)))
         if (tileEncoding not in (TILE_FORMAT_PIL, TILE_FORMAT_NUMPY) and
                 numpyAllowed != 'always' and tileEncoding == self.encoding and
-                not isEdge and (not applyStyle or not getattr(self, 'style', None))):
+                not isEdge and (not applyStyle or not hasStyle)):
             return tile
         mode = None
-        if (numpyAllowed == 'always' or tileEncoding == TILE_FORMAT_NUMPY or (applyStyle and (
-                (getattr(self, 'style', None) or hasattr(self, '_iccprofiles')) and
-                getattr(self, 'style', None) != {'icc': False})) or
-                isEdge):
+        if (numpyAllowed == 'always' or tileEncoding == TILE_FORMAT_NUMPY or
+                (applyStyle and hasStyle) or isEdge):
             tile, mode = self._outputTileNumpyStyle(
                 tile, applyStyle, x, y, z, self._getFrame(**kwargs))
         if isEdge: