+++ CHANGELOG.md
@@ -3,13 +3,14 @@
 ## 1.20.0
 
 ### Features
-- ICC color profile support ([#1037](../../pull/1037))
+- ICC color profile support ([#1037](../../pull/1037), [#1043](../../pull/1043), [#1046](../../pull/1046), [#1048](../../pull/1048), [#1052](../../pull/1052))
 
 ### Improvements
-- Speed up generating tiles for some multi source files ([#1035](../../pull/1035))
+- Speed up generating tiles for some multi source files ([#1035](../../pull/1035), [#1047](../../pull/1047))
 - Render item lists faster ([#1036](../../pull/1036))
 - Reduce bioformats source memory usage ([#1038](../../pull/1038))
 - Better pick the largest image from bioformats ([#1039](../../pull/1039), [#1040](../../pull/1040))
+- Cache histogram thresholds ([#1042](../../pull/1042))
 
 ## 1.19.3